requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
pytesseract==0.3.8
Pillow==8.2.0
azure-ai-textanalytics==5.1.0b7
azure-storage-blob==12.8.1